refactor: Split integration suite into flat per-capability cases
Splits tests/integration/suite.ts (3292 lines, 117 cases) into
tests/integration/cases/*.cases.ts, grouped flat by capability
(registration, tools, actors, apps, tasks, storage, payments), per
the plan agreed in #776 (flat structure, no folder-per-tool mirror of
src/). Every register*Cases(ctx) is a mechanical move — same test
names, bodies, and transport gates; suite.ts now only wires the three
transport dimensions and calls the register functions. Verified via
AST title diff: 117 titles before, 117 after, zero difference.

Adds src/test_kit/ (published behind the "./test-kit" package export,
vitest optional peerDependency) so apify-mcp-server-internal can
import and run a curated critical: true subset against its own live
deploy instead of hand-duplicating assertions. Wires two scenarios
(tests/integration/cases/shared_scenarios.ts) through it as a working
example; growing that subset is a per-PR judgment call, left for
follow-up.

Updates AGENTS.md and DEVELOPMENT.md per #776's own requirement to
document the new structure when the first capability module lands.

Fixes #776

AI use disclosed: mechanical statement extraction (line-range slicing
driven by a TypeScript AST parse) was scripted; grouping decisions,
the test-kit design, and all verification were reviewed by hand.
